Breaking out Game checks rom CnCNet Lobby
* New Game Check class with removal of thread.sleep
