+++ DXMainClient/DXMainClient.csproj
@@ -315,6 +315,7 @@
     <Compile Include="DXGUI\Multiplayer\CnCNet\PrivateMessagingPanel.cs" />
     <Compile Include="DXGUI\Multiplayer\CnCNet\PrivateMessagingWindow.cs" />
     <Compile Include="Online\ChannelUser.cs" />
+    <Compile Include="Online\CnCNetGameCheck.cs" />
     <Compile Include="Online\EventArguments\ChannelCTCPEventArgs.cs" />
     <Compile Include="Online\EventArguments\PrivateMessageEventArgs.cs" />
     <Compile Include="Online\FileHashCalculator.cs" />
